Created Tracker data model to store database model. Added DatabaseManager class, moved persistentContainer property and saveContext method in this class. refactored AppDelegate and SceneDelegate classes.
